+++ coursework/src/Factories.h
@@ -11,7 +11,7 @@
 //object factories
 SatelliteShared createSatellite(
 	const ModelShared& model
-	, float mass
+	, double mass
 	, const Vec3& color = Vec3(1.0f)
 	, const Mat4& mat = Mat4(1.0f)
 	, const Vec3& pos = Vec3()
@@ -22,7 +22,7 @@ SatelliteShared createSatellite(
 
 SatelliteShared createChaser(
 	const ModelShared& model
-	, float mass
+	, double mass
 	, const Vec3& color = Vec3(1.0f)
 	, const Mat4& mat = Mat4(1.0f)
 	, const Vec3& pos = Vec3()
@@ -34,7 +34,7 @@ SatelliteShared createChaser(
 
 PlanetShared createPlanet(
 	const ModelShared& model
-	, float mass
+	, double mass
 	, const Mat4& mat = Mat4(1.0f)
 	, const Vec3& pos = Vec3()
 	, const Vec3& vel = Vec3()
