+++ coursework/src/Objects/Satellite/Rendezvous/LambertSolver.h
@@ -2,13 +2,13 @@
 
 #include <Common.h>
 
-float solveLambertProblem(
+double solveLambertProblem(
 	const Vec3& rv1
-	, float t1
+	, double t1
 	, const Vec3& rv2
-	, float t2
-	, float mu
-	, float x0
-	, float eps = 1e-10f
+	, double t2
+	, double mu
+	, double x0
+	, double eps = 1e-10f
 	, int limit = 10
 );
